requirements.txt
anyio==3.7.0
asgiref==3.7.2
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
decorator==5.1.1
Django==4.2.2
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
infinity==1.5
install==1.3.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-telegram-bot==20.3
python-telegram-bot-pagination==0.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
typing_extensions==4.6.3
validators==0.20.0
Werkzeug==2.3.6
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
WTForms-Ext==0.5
WTForms-SQLAlchemy==0.3
